ci(sqlite): run the full suite on the default backend; split conftest (#220)
Three things #220 asked for that were still open after the parity fixes.

1. CI actually gates SQLite now. The job ran ONE file
   (test_sqlite_backend.py) behind a comment deferring the rest to a
   "full-parity effort" that named no issue. That is precisely how the
   three defects fixed in 17be6b5 reached the plugin's DEFAULT backend
   without CI noticing. It now runs the full suite, with
   CORTEX_MEMORY_STORE_BACKEND=sqlite set explicitly rather than inferred
   from the absence of a PG service container, so a future runner that
   happens to have PostgreSQL reachable cannot silently turn this into a
   second PostgreSQL run.

2. conftest.py was 606 lines against a 500-line cap (§4.1) — a defect in
   material this work touched, so it is fixed here rather than noted.
   Between-test cleanup and singleton reset moved verbatim to
   tests_py/_store_cleanup.py (conftest 437, new module 196). The one
   behavioural-adjacent change is the SQLite path binding: the moved code
   reads CORTEX_MEMORY_DB_PATH instead of conftest's _ISOLATED_SQLITE_PATH
   global, which removes an import cycle. It is the same value, and the
   import sits BELOW _redirect_real_data_roots() because that call is what
   sets the variable — the #219 ordering constraint, restated at the site.

3. test_I2_canonical_writer pins heat writers by LINE NUMBER and broke
   twice in this session from edits that only shifted lines (an import,
   then a docstring). Pins realigned to 530/560/624, each verified by
   reading the line. The fragility is real and worth replacing with
   content-based anchors; that is a separate change, not a silent
   deferral.

Verification (full suites, both backends):
  SQLITE_EXIT=0  5639 passed, 96 skipped
  PG_EXIT=0      5723 passed, 12 skipped
  ruff check + format clean (931 files); doc claims OK at 5735
  real ~/.claude store checksummed before/after: untouched
